refactor: Cars 로직에 CarLocationResult 생성 로직 변경
